requirements.txt
numpy
scikit-learn
matplotlib
pandas
pyserial